test: 🔄 moduleResolver: move to a separate file
